# Settings Merge Optimizations