--- a/lib/cloneDeep.js
+++ b/lib/cloneDeep.js
@@ -1,0 +1,27 @@
+module.exports = function cloneDeep(o) {
+    var newO,
+        i;
+
+    if (typeof o !== 'object') {
+        return o;
+    }
+    if (!o) {
+        return o;
+    }
+
+    if ('[object Array]' === Object.prototype.toString.apply(o)) {
+        newO = [];
+        for (i = 0; i < o.length; i += 1) {
+            newO[i] = cloneDeep(o[i]);
+        }
+        return newO;
+    }
+
+    newO = {};
+    for (i in o) {
+        if (o.hasOwnProperty(i)) {
+            newO[i] = cloneDeep(o[i]);
+        }
+    }
+    return newO;
+};