--- a/lib/isA.js
+++ b/lib/isA.js
@@ -1,0 +1,3 @@
+module.exports = function isA(item, constructor) {
+    return item && (item.constructor === constructor);
+};