--- a/lib/isIterable.js
+++ b/lib/isIterable.js
@@ -1,0 +1,5 @@
+var isA = require('./isA');
+
+module.exports = function isIterable(item) {
+    return isA(item, Object) || isA(item, Array);
+};