Change merge to ensure new object does not reference old object so that clone does not need to be called as much during read.

```
+---------------------------------------+-----------------+-----------------+
|                                       │ YCB             │ New             |
+---------------------------------------+-----------------+-----------------+
| Base                                  │ 156,017 ops/sec │ 225,027 ops/sec |
+---------------------------------------+-----------------+-----------------+
| y20 (large config / large dimensions) │ 14,017 ops/sec  │ 16,776 ops/sec  |
+---------------------------------------+-----------------+-----------------+
| y20 (large config / small dimensions) │ 16,695 ops/sec  │ 19,972 ops/sec  |
+---------------------------------------+-----------------+-----------------+
| y20 (small config / large dimensions) │ 59,729 ops/sec  │ 66,058 ops/sec  |
+---------------------------------------+-----------------+-----------------+
| y20 (small config / small dimensions) │ 113,577 ops/sec │ 127,480 ops/sec |
+---------------------------------------+-----------------+-----------------+
```

In doing this optimization, `mergeDeep` needed access to `cloneDeep` so it was moved outside of the class. In order to keep the unit tests and add new ones I moved `mergeDeep` and `cloneDeep` to their own files (along with the other shared utility functions).